sdcc: document register map, fix naming, improve code quality
- Document full MCI + SDHCI register maps in sdcc_regs.c header
- Add register offset annotations to firehose.h declarations
- Rename mmc_read_ext_csd → mmc_alloc_handle (function allocates
  partition table entry, not reading EXT_CSD)
- Rename DEV_EXT_CSD_PTR → DEV_HOTPLUG_DESC (dev[0x24] points to
  hotplug descriptor, not raw EXT_CSD data)
- Fix misleading (short) cast → (int) in sdcc_pre_write_setup
- Fix wrong comment in card_init.c slot context layout

+++ src_minimal/sdcc_helpers.c
@@ -608,11 +608,11 @@ int sdcc_pre_write_setup(undefined4 *dev, int is_reliable, int num_blocks)
   else {
     iVar2 = dev[0x21] * iVar2;
   }
-  *(int *)(DAT_0804e2c8[(short)uVar1] + 0x24) = iVar2;
+  *(int *)(DAT_0804e2c8[(int)uVar1] + 0x24) = iVar2;
   if (dev[0x16] == 1) {
     uVar3 = dev[9] & 0xffff;
   }
-  *(uint *)(DAT_0804e2c8[(short)uVar1] + 0x28) = uVar3 * num_blocks;
+  *(uint *)(DAT_0804e2c8[(int)uVar1] + 0x28) = uVar3 * num_blocks;
   return uVar3 * num_blocks;
 }
